Move shared widgets to transforms/components
