--- enterprise/frontend/src/metabase-enterprise/transforms/pages/RunListPage/FilterList/constants.ts
@@ -1,2 +1,0 @@
-export const MIN_WIDTH = 300;
-export const MAX_HEIGHT = 400;
